pyproject.toml
[tool.poetry]
name = "agconnect"
version = "0.1.0"
description = "Agriculture multi-agent"
authors = ["IDOEFRAIM <idoefraimtanka@gmail.com>"]
license = "MIT"
readme = "README.md"

# 👇 Add this line to tell Poetry where your code is
packages = [
    { include = "agents" },
    {include="rag"},
    {include="Tools"},
    {include="services"}
]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
numpy = "^1.26"
PyPDF2 = "^3.0"
playwright = "^1.45"
PyMuPDF = "^1.24"
chromadb = {extras = ["duckdb"], version = "^0.5"}
sentence-transformers = "^3.0"
tqdm = "^4.66"
pydantic = "^2.0"
httpx = "^0.27"
shapely = "^2.0"
pillow = "^10.0"
loguru = "^0.7"
aiofiles = "^23.1"
pdfminer-six = "^20221105"
pytest = "^7.0"
fastapi = "^0.115"
aiohttp = "^3.9"
torch = "^2.2"
pymilvus = "^2.4"
aioredis = "^2.0"
srsly = "^2.5.0"
spacy = "^3.7"
langgraph = "^1.0.4"
langchain = "^1.0.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"